Refactor common-code out of wiring_diagram.
