+++ spinner/diagrams/machine_map.py
