+++ tests/scripts/test_arguments.py
@@ -15,6 +15,93 @@ from spinner import utils
 from example_cabinet_params import board_wire_offset_fields, unique
 
 
+class TestCabinetAction(object):
+	
+	@pytest.fixture
+	def parser(self):
+		parser = ArgumentParser()
+		
+		parser.add_argument("--cabinet-only", action=arguments.CabinetAction(1))
+		parser.add_argument("--frame-only", action=arguments.CabinetAction(2))
+		parser.add_argument("--board-only", action=arguments.CabinetAction(3))
+		parser.add_argument("--full-spec", action=arguments.CabinetAction(4))
+		parser.add_argument("--multiple", action=arguments.CabinetAction(4, append=True))
+		
+		return parser
+	
+	
+	@pytest.mark.parametrize("argstring", [# No arguments
+	                                       "--cabinet-only",
+	                                       "--frame-only",
+	                                       "--board-only",
+	                                       "--full-spec",
+	                                       "--multiple",
+	                                       "--multiple --multiple",
+	                                       # Too many arguments
+	                                       "--cabinet-only 1 2",
+	                                       "--frame-only 1 2 3",
+	                                       "--board-only 1 2 3 north",
+	                                       "--full-spec 1 2 3 north bad",
+	                                       "--multiple 1 2 3 north bad",
+	                                       "--multiple 1 2 3 north bad --multiple 1 2 3 north bad",
+	                                       # Negative
+	                                       "--cabinet-only -1",
+	                                       "--frame-only 1 -2",
+	                                       "--board-only 1 2 -3 north",
+	                                       "--full-spec 1 2 -3 north",
+	                                       "--multiple 1 2 -3 north",
+	                                       "--multiple 1 2 -3 north --multiple 1 2 -3 north",
+	                                       # Non-number
+	                                       "--cabinet-only bad",
+	                                       "--frame-only 1 bad",
+	                                       "--board-only 1 2 bad north",
+	                                       "--full-spec 1 2 bad north",
+	                                       "--multiple 1 2 bad north",
+	                                       "--multiple 1 2 bad north --multiple 1 2 bad north",
+	                                       # Non-direction
+	                                       "--full-spec 1 2 3 bad",
+	                                       "--multiple 1 2 3 bad",
+	                                       "--multiple 1 2 3 bad --multiple 1 2 3 bad",
+	                                      ])
+	def test_bad(self, parser, argstring):
+		with pytest.raises(SystemExit):
+			parser.parse_args(argstring.split())
+	
+	
+	@pytest.mark.parametrize("argstring,full_spec",
+	                         [# Just cabinet
+	                          ("--full-spec 0", (0, )),
+	                          ("--full-spec 1", (1, )),
+	                          # Cabinet and frame
+	                          ("--full-spec 0 0", (0, 0)),
+	                          ("--full-spec 1 2", (1, 2)),
+	                          # Cabinet, frame and board
+	                          ("--full-spec 0 0 0", (0, 0, 0)),
+	                          ("--full-spec 1 2 3", (1, 2, 3)),
+	                          # Cabinet, frame, board and socket
+	                          ("--full-spec 0 0 0 north",
+	                           (0, 0, 0, Direction.north)),
+	                          ("--full-spec 1 2 3 north-east",
+	                           (1, 2, 3, Direction.north_east)),
+	                         ])
+	def test_full_spec(self, parser, argstring, full_spec):
+		args = parser.parse_args(argstring.split())
+		assert args.full_spec == full_spec
+	
+	
+	@pytest.mark.parametrize("argstring,multiple",
+	                         [# None
+	                          ("", None),
+	                          # One
+	                          ("--multiple 1", [(1, )]),
+	                          # Several
+	                          ("--multiple 1 --multiple 2 3", [(1, ), (2, 3)]),
+	                         ])
+	def test_multiple(self, parser, argstring, multiple):
+		args = parser.parse_args(argstring.split())
+		assert args.multiple == multiple
+
+
 @pytest.mark.parametrize("argstring",
                          ["",  # Requires -n or -t
                           "-n 12 -t 8 8",  # ...but not both
@@ -255,3 +342,103 @@ def test_get_histogram_from_args_bad(argstring):
 		args = parser.parse_args(argstring.split())
 		arguments.get_histogram_from_args(parser, args)
 
+
+@pytest.mark.parametrize("argstring,aspect_ratio,to_check",
+                         [# Passes through the filename...
+                          ("/super/happy/smiley.png", 0.5,
+                           {"output_filename": "/super/happy/smiley.png"}),
+                          # File type detection
+                          ("out.png", 0.5, {"file_type": "png"}),
+                          ("out.pNg", 0.5, {"file_type": "png"}),
+                          ("out.PNG", 0.5, {"file_type": "png"}),
+                          ("out.pdf.png", 0.5, {"file_type": "png"}),
+                          ("out.pdf", 0.5, {"file_type": "pdf"}),
+                          ("out.pDf", 0.5, {"file_type": "pdf"}),
+                          ("out.PDF", 0.5, {"file_type": "pdf"}),
+                          ("out.png.pdf", 0.5, {"file_type": "pdf"}),
+                          # Manual image sizes (PNGs should be integers, PDFs
+                          # should be floats)
+                          ("out.png 10.5 100.5", 0.5, {"image_width":10,
+                                                       "image_height":100}),
+                          ("out.png 100.5 10.5", 0.5, {"image_width":100,
+                                                       "image_height":10}),
+                          ("out.pdf 10.5 100.5", 0.5, {"image_width":10.5,
+                                                       "image_height":100.5}),
+                          ("out.pdf 100.5 10.5", 0.5, {"image_width":100.5,
+                                                       "image_height":10.5}),
+                          # Semi-automatic for tall images
+                          ("out.pdf 1000.5", 0.5, {"image_height":1000.5,
+                                                   "aspect_ratio":0.5}),
+                          ("out.png 1000.5", 0.5, {"image_height":1000,
+                                                   "aspect_ratio":0.5}),
+                          # Semi-automatic but for wider images
+                          ("out.pdf 1000.5", 1.5, {"image_width":1000.5,
+                                                    "aspect_ratio":1.5}),
+                          ("out.png 1000.5", 1.5, {"image_width":1000,
+                                                    "aspect_ratio":1.5}),
+                          # Fully-automatic for tall images
+                          ("out.png", 0.5, {"image_height":1000,
+                                            "aspect_ratio":0.5}),
+                          ("out.pdf", 0.5, {"image_height":280.0,
+                                            "aspect_ratio":0.5}),
+                          # Fully-automatic for wide images
+                          ("out.png", 1.5, {"image_width":1000,
+                                             "aspect_ratio":1.5}),
+                          ("out.pdf", 1.5, {"image_width":280.0,
+                                             "aspect_ratio":1.5}),
+                         ])
+def test_get_arguments(argstring, aspect_ratio, to_check):
+	parser = ArgumentParser()
+	arguments.add_image_args(parser)
+	
+	args = parser.parse_args(argstring.split())
+	output_filename, file_type, image_width, image_height =\
+		arguments.get_image_from_args(parser, args, aspect_ratio)
+	
+	if "output_filename" in to_check:
+		assert output_filename == to_check.pop("output_filename")
+	
+	if "file_type" in to_check:
+		assert file_type == to_check.pop("file_type")
+	
+	if "image_height" in to_check:
+		assert image_height == to_check.pop("image_height")
+	if "image_width" in to_check:
+		assert image_width == to_check.pop("image_width")
+	
+	if "aspect_ratio" in to_check:
+		ref_ratio = to_check.pop("aspect_ratio")
+		ratio = image_width / float(image_height)
+		
+		# Should be pretty close (slack allowed for rounding to pixels in PNG)
+		assert abs(ratio - ref_ratio) < 0.01
+	
+	# Make sure none of the test-cases define something to test which isn't
+	# tested...
+	assert len(to_check) == 0
+
+
+@pytest.mark.parametrize("argstring",
+                         [# Missing filename
+                          "",
+                          # Missing/Unknown file extension
+                          "out",
+                          "out.gif",
+                          "out.png.gif",
+                          # Invalid output sizes
+                          "out.png 0.5",
+                          "out.png 0.5 0.5",
+                          "out.png 10 0.5",
+                          "out.png 0.5 10",
+                          "out.pdf 0",
+                          "out.pdf 0 0",
+                          "out.pdf 1 0",
+                          "out.pdf 0 1",
+                         ])
+def test_get_arguments_bad(argstring):
+	parser = ArgumentParser()
+	arguments.add_image_args(parser)
+	
+	with pytest.raises(SystemExit):
+		args = parser.parse_args(argstring.split())
+		arguments.get_image_from_args(parser, args)
